requirements.txt
asgiref==3.4.0
certifi==2021.5.30
Django==3.2.4
djangorestframework==3.12.4
pip==21.1.3
psycopg2==2.8.6
pytz==2021.1
setuptools==52.0.0.post20210125
sqlparse==0.4.1
typing-extensions==3.10.0.0
wheel==0.36.2
